third_party/WebKit/LayoutTests/inspector/filtered-item-selection-dialog-rendering-expected.txt

Verifies that SelectUISourceCodeDialog rendering works properly.


Running: testRenderingInNameOnly
filter query: 12
helloWorld[1][2].js
test/helloWorld12.js
--------------------
are-shown.js
test/some/very-long-url/which/usually/breaks-rendering/due-to/trancation/so/that/the-path-is-cut-ap…
--------------------

Running: testRenderingInPathAndName
filter query: te12
helloWorld12.js
[t][e]st/helloWorld[1][2].js
--------------------
are-shown.js
test/some/very-long-url/which/usually/breaks-rendering/due-to/trancation/so/that/the-path-is-cut-ap…
--------------------

Running: testRenderingInNameInTruncatedPath
filter query: shown.js
helloWorld12[.][j][s]
test/helloWorld12.js
--------------------
are-[s][h][o][w][n][.][j][s]
test/some/very-long-url/which/usually/breaks-rendering/due-to/trancation/so/that/the-path-is-cut-ap…
--------------------

Running: testRenderingInTruncatedPath
filter query: usually-shown.js
helloWorld12[.][j][s]
test/helloWorld12.js
--------------------
are-shown.js
test/some/very-long-url/which/[u][s][u][a][l][l][y]/breaks-rendering/due-to/trancation/so/that/the-path-is-cut-ap…
--------------------
